+++ Changelog.md
@@ -160,13 +160,13 @@ like knowing which weapon is better or whether an attachment will be useful to y
 - Updated dependencies:
   - Updated to [gradle-buildconfig-plugin](https://github.com/gmazzo/gradle-buildconfig-plugin) 5.7.1
   - Updated to [gradle-idea-ext-plugin](https://github.com/JetBrains/gradle-idea-ext-plugin) 1.3
-  - Updated to [io.freefair.lombok](https://plugins.gradle.org/plugin/io.freefair.lombok) 8.14.3
-  - Updated to [RetroFuturaGradle](https://github.com/GTNewHorizons/RetroFuturaGradle) 1.4.9
+  - Updated to [io.freefair.lombok](https://plugins.gradle.org/plugin/io.freefair.lombok) 9.1.0
+  - Updated to [RetroFuturaGradle](https://github.com/GTNewHorizons/RetroFuturaGradle) 2.0.0
   - Updated to [foojay-resolver](https://github.com/gradle/foojay-toolchains) 1.0.0
   - Updated to [GroovyScript](https://www.curseforge.com/minecraft/mc-mods/groovyscript) 1.3.1
   - Updated to [MixinBooter](https://www.curseforge.com/minecraft/mc-mods/mixin-booter) 10.7
   - Updated to [Red Core](https://www.curseforge.com/minecraft/mc-mods/red-core) 0.6
-  - Updated to [Gradle](https://gradle.org) 8.14.3
+  - Updated to [Gradle](https://gradle.org) 9.2.0
 
 ## Modern Warfare Cubed Version 0.1.9 Changelog - 2024-11-25